test(jsx): split module emission tests out of compile.rs
